+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Mapbox welcomes participation and contributions from everyone.
 
 ## main
 
+* Expose a method to remove tile region with a completion: `TileStore.removeTileRegion(forId:completion:)`.
+
 ## 10.16.0-rc.1 - August 23, 2023
 
 * Add privacy policy attribution dialog action.